+++ src/app/service/poap/ClaimPOAP.ts
@@ -1,11 +1,10 @@
 import { CommandContext, EmbedField } from 'slash-create';
-import { GuildMember, MessageEmbed } from 'discord.js';
+import { GuildMember } from 'discord.js';
 import { Collection, Cursor, Db } from 'mongodb';
 import MongoDbUtils from '../../utils/MongoDbUtils';
 import constants from '../constants/constants';
 import { POAPUnclaimedParticipants } from '../../types/poap/POAPUnclaimedParticipants';
 import Log from '../../utils/Log';
-import { MessageEmbedOptions } from 'slash-create/lib/structures/message';
 import { POAPTwitterUnclaimedParticipants } from '../../types/poap/POAPTwitterUnclaimedParticipants';
 import VerifyTwitter, { VerifiedTwitter } from '../account/VerifyTwitter';
 
@@ -58,9 +57,9 @@ const ClaimPOAP = async (ctx: CommandContext, platform: string, guildMember?: Gu
 		unclaimedParticipantsCollection.deleteMany({
 			discordUserId: ctx.user.id,
 		});
-	}, 1000);
+	}, 600000);
 	
-	Log.debug('POAP claimed and deleted from db');
+	Log.debug('POAP claimed');
 };
 
 const claimPOAPForTwitter = async (ctx: CommandContext, guildMember: GuildMember) => {
@@ -83,51 +82,34 @@ const claimPOAPForTwitter = async (ctx: CommandContext, guildMember: GuildMember
 		return;
 	}
 	Log.debug('POAP found');
-
-	if (guildMember != null) {
-		Log.debug('attempting to send message from channel!');
-		const embedMessageList: MessageEmbed[] = await unclaimedParticipants.map((doc: POAPTwitterUnclaimedParticipants) => {
-			return new MessageEmbed({
-				title: 'POAP link',
-				description: 'Thank you for participating in the community event!',
-				fields: [
-					{ name: 'Discord', value: `${doc.discordServerName}`, inline: false },
-					{ name: 'Event', value: `${doc.event}`, inline: true },
-					{ name: 'Claim Link', value: `${doc.poapLink}`, inline: true },
-				],
-			});
-		}).toArray();
-
-		await guildMember.send({
-			embeds: embedMessageList,
-		});
-		await guildMember.send({ content: 'Oh goody! We found some POAPS!' });
-		await ctx.send({ content: 'Sent you a DM!' });
-	} else {
-		Log.debug('attempting to send message from DM!');
-		const embedMessageList: MessageEmbedOptions[] = await unclaimedParticipants.map((doc: POAPTwitterUnclaimedParticipants) => {
-			return {
-				title: 'POAP link',
-				description: 'Thank you for participating in the community event!',
-				fields: [
-					{ name: 'Discord', value: `${doc.discordServerName}`, inline: false },
-					{ name: 'Event', value: `${doc.event}`, inline: true },
-					{ name: 'Claim Link', value: `${doc.poapLink}`, inline: true },
-				],
-			} as MessageEmbedOptions;
-		}).toArray();
-		await ctx.send({
-			embeds: embedMessageList,
-		});
-		await ctx.send({ content: 'Oh goody! We found some POAPS!' });
-	}
-	Log.debug('message sent to user!');
-
-	await unclaimedParticipantsCollection.deleteMany({
-		twitterUserId: verifiedTwitter.twitterUser.id_str,
+	
+	const fieldsList: EmbedField[] = await unclaimedParticipants.map((doc: POAPTwitterUnclaimedParticipants) => {
+		return ({
+			name: 'Claim Link',
+			value: `${doc.poapLink}`,
+			inline: false,
+		} as EmbedField);
+	}).toArray();
+	
+	await ctx.send({
+		embeds: [{
+			title: 'POAP link',
+			description: 'Thank you for participating in the community event!',
+			fields: fieldsList,
+		}],
+		ephemeral: true,
 	});
 	
-	Log.debug('POAP claimed and deleted from twitter db');
+	Log.debug('message sent to user!');
+
+	setTimeout(() => {
+		Log.debug('deleted poaps from claim twitter');
+		unclaimedParticipantsCollection.deleteMany({
+			twitterUserId: verifiedTwitter.twitterUser.id_str,
+		});
+	}, 600000);
+	
+	Log.debug('POAP claimed.');
 };
 
 export default ClaimPOAP;