+++ src/app/utils/ServiceUtils.ts
@@ -7,22 +7,19 @@ import {
 	Guild,
 	GuildMember,
 	Permissions,
-	Role,
-	RoleManager,
 	Snowflake,
 	StageChannel,
 	User,
 	VoiceChannel,
 } from 'discord.js';
-import { CommandContext } from 'slash-create';
 import client from '../app';
 import ValidationError from '../errors/ValidationError';
-import roleIDs from '../service/constants/roleIds';
-import discordServerIds from '../service/constants/discordServerIds';
 import Log, { LogUtils } from './Log';
 import { stringify } from 'csv-stringify/sync';
 import { parse } from 'csv-parse/sync';
 import { POAPFileParticipant, TwitterPOAPFileParticipant } from './POAPUtils';
+import { ButtonStyle, CommandContext, ComponentType } from 'slash-create';
+import { ComponentActionRow } from 'slash-create/lib/constants';
 
 const ServiceUtils = {
 	async getGuildAndMember(ctx: CommandContext): Promise<{ guild: Guild, guildMember: GuildMember }> {
@@ -42,24 +39,6 @@ const ServiceUtils = {
 		const guildMembers = await guild.members.fetch();
 		return guildMembers.filter(member => {
 			return ServiceUtils.hasSomeRole(member, roles);
-		});
-	},
-
-	getGuestRole(roles: RoleManager): Role {
-		return roles.cache.find((role) => {
-			return role.id === roleIDs.guestPass;
-		});
-	},
-
-	getAFKRole(roles: RoleManager): Role {
-		return roles.cache.find((role) => {
-			return role.id === roleIDs.AFK;
-		});
-	},
-	
-	getRole(roleManager: RoleManager, roleName: string): Role {
-		return roleManager.cache.find(role => {
-			return role.name === roleName;
 		});
 	},
 
@@ -83,42 +62,6 @@ const ServiceUtils = {
 	isDiscordServerManager(guildMember: GuildMember): boolean {
 		return guildMember.permissions.has(Permissions.FLAGS.MANAGE_GUILD);
 	},
-
-	isAnyLevel(guildMember: GuildMember): boolean {
-		return ServiceUtils.hasSomeRole(guildMember, [
-			roleIDs.level1,
-			roleIDs.level2,
-			roleIDs.level3,
-			roleIDs.level4,
-		]);
-	},
-
-	isAtLeastLevel1(guildMember: GuildMember): boolean {
-		return ServiceUtils.hasSomeRole(guildMember, [
-			roleIDs.level1,
-			roleIDs.level2,
-			roleIDs.level3,
-			roleIDs.level4,
-			roleIDs.admin,
-			roleIDs.genesisSquad,
-		]);
-	},
-
-	isAtLeastLevel2(guildMember: GuildMember): boolean {
-		return ServiceUtils.hasSomeRole(guildMember, [
-			roleIDs.level2,
-			roleIDs.level3,
-			roleIDs.level4,
-			roleIDs.admin,
-			roleIDs.genesisSquad,
-		]);
-	},
-	
-	validateLevel2AboveMembers(guildMember: GuildMember): void {
-		if (!(ServiceUtils.isAtLeastLevel2(guildMember))) {
-			throw new ValidationError('Must be `level 2` or above member.');
-		}
-	},
 	
 	formatDisplayDate(dateIso: string): string {
 		const options: Intl.DateTimeFormatOptions = {
@@ -128,13 +71,6 @@ const ServiceUtils = {
 			year: 'numeric',
 		};
 		return (new Date(dateIso)).toLocaleString('en-US', options);
-	},
-	
-	isBanklessDAO(guild: Guild): boolean {
-		if (guild == null || guild.id == null) {
-			return false;
-		}
-		return guild.id == discordServerIds.banklessDAO || guild.id == discordServerIds.discordBotGarage;
 	},
 	
 	getAllVoiceChannels(guildMember: GuildMember): Collection<string, VoiceChannel | StageChannel> {
@@ -163,7 +99,7 @@ const ServiceUtils = {
 			await guildMember.send({ content: message });
 		} catch (e) {
 			LogUtils.logError('DM is turned off', e);
-			throw new ValidationError('I\'m trying to send you a DM... Can you try turning DMs on?');
+			throw new ValidationError('Please turn on direct messages.');
 		}
 	},
 	
@@ -192,6 +128,23 @@ const ServiceUtils = {
 		});
 		Log.debug('done parsing csv file');
 		return records;
+	},
+	
+	sendOutErrorMessage: async (ctx: CommandContext): Promise<any> => {
+		const row: ComponentActionRow = {
+			type: ComponentType.ACTION_ROW,
+			components: [{
+				type: ComponentType.BUTTON,
+				style: ButtonStyle.LINK,
+				label: 'Support',
+				url: 'https://discord.gg/NRj43H83nJ',
+			}],
+		};
+		await ctx.send({
+			content: 'Something is not working. Please reach out to us and a support member will happily assist!',
+			ephemeral: true,
+			components: [row],
+		});
 	},
 };